Add custom email authentication backend and update settings

+++ filesystem/views.py
@@ -8,50 +8,47 @@ from django.http import HttpResponseForbidden, FileResponse
 from .models import FileModels
 from django.core.mail import EmailMessage
 from django.conf import settings
-from django.utils. decorators import method_decorator
+from django.utils.decorators import method_decorator
 from django.views.generic import ListView, DetailView
 import os
 from django.core.paginator import Paginator
 
-
 @login_required
 def upload_file(request):
-    user= CustomUser.objects.get(pk=request.user.id)
+    user = CustomUser.objects.get(pk=request.user.id)
     if user.is_superuser:
-        if request.method =="POST":
+        if request.method == "POST":
             form = FileForm(request.POST, request.FILES)
             if form.is_valid():
                 file = form.save()
                 if file.file.name.lower().endswith('.pdf'):
-                    file = form.save()
                     file.user = request.user
                     file_path = file.file.path
                     thumbnail = generate_thumbnail(file_path)
-
                     file.thumbnail.save(f'{file.title}_thumbnail.jpg', ContentFile(thumbnail))
                 else:
                     form.save()
-                    print('user: ' + file.user)
                 return redirect('filesystem:upload_list')
         else:
             form = FileForm()
-
-        return render(request, 'filesystem/upload_file.html', {'form': form}) 
+        return render(request, 'filesystem/upload_file.html', {'form': form})
     else:
         return HttpResponseForbidden('<h1>You are not authorised to view this page</h1>')
 
-
-#delete method for deleting files
+@login_required
 def delete_upload_file(request, pk):
-    file =FileModels.objects.get(pk=pk)
-    file.delete()
-    file_path = os.path.join(settings.MEDIA_ROOT, str(file.file))
-    if os.path.exists(file_path):
+    user = CustomUser.objects.get(pk=request.user.id)
+    if user.is_superuser:
+        file = FileModels.objects.get(pk=pk)
+        file.delete()
+        file_path = os.path.join(settings.MEDIA_ROOT, str(file.file))
+        if os.path.exists(file_path):
             os.remove(file_path)
-    return redirect('filesystem:upload_list')
+        return redirect('filesystem:upload_list')
+    else:
+        return HttpResponseForbidden('<h1>You are not authorised to view this page</h1>')
 
-
-# @login_required
+@login_required
 def download_file(request, file_id):
     file = FileModels.objects.get(pk=file_id)
     response = FileResponse(file.file, as_attachment=True)
@@ -59,9 +56,7 @@ def download_file(request, file_id):
     file.save()
     return response
 
-
-
-# @login_required
+@login_required
 def send_file_email(request, file_id):
     file = get_object_or_404(FileModels, pk=file_id)
     if request.method == 'POST':
@@ -70,13 +65,13 @@ def send_file_email(request, file_id):
             recipient_email = form.cleaned_data['recipient_email']
             subject = form.cleaned_data['subject']
             message = form.cleaned_data['message']
-            email= EmailMessage(
-                subject, 
-                message, 
-                settings.EMAIL_HOST_USER, 
-                [recipient_email], 
+            email = EmailMessage(
+                subject,
+                message,
+                settings.EMAIL_HOST_USER,
+                [recipient_email],
                 reply_to=['another@example.com']
-                )
+            )
             email.attach_file(file.file.path)
             email.send()
             file.emails_sent += 1
@@ -85,10 +80,8 @@ def send_file_email(request, file_id):
     else:
         form = SendFileForm()
     return render(request, 'filesystem/send_file.html', {'form': form, 'file': file})
-    
 
-
-# @method_decorator(login_required, name='dispatch')
+@method_decorator(login_required, name='dispatch')
 class FileListView(ListView):
     model = FileModels
     template_name = 'filesystem/upload_list.html'
@@ -96,62 +89,48 @@ class FileListView(ListView):
     ordering = ['title']
     paginate_by = 10
 
-
-
-
-# @method_decorator(login_required, name='dispatch')
+@method_decorator(login_required, name='dispatch')
 class FileDetailView(DetailView):
     model = FileModels
     template_name = 'filesystem/file_detail.html'
 
     def get_context_data(self, **kwargs):
-        parameter= super().get_context_data(**kwargs)
+        parameter = super().get_context_data(**kwargs)
         parameter['file'] = self.get_object()
         return parameter
-    
-
 
 @login_required
 def log(request):
     user = CustomUser.objects.get(pk=request.user.pk)
     if user.is_superuser:
         files = FileModels.objects.all()
-        #creating of a pagination table
         paginator = Paginator(files, 10)
         page_number = request.GET.get('page')
-        page_obj =paginator.get_page(page_number)
-        return render(request, 'filesystem/logs.html', {'files': files, 'page_obj':page_obj})
+        page_obj = paginator.get_page(page_number)
+        return render(request, 'filesystem/logs.html', {'files': files, 'page_obj': page_obj})
     else:
-        return HttpResponseForbidden('<h1>You are not authorised to view thsi page</h1>')
-    
+        return HttpResponseForbidden('<h1>You are not authorised to view this page</h1>')
 
-
-# @login_required
+@login_required
 def search_view(request):
     query = request.GET.get('q')
     if query:
-        files = FileModels.objects.filter(title__icontains = query)
+        files = FileModels.objects.filter(title__icontains=query)
     else:
         files = []
     return render(request, 'filesystem/search.html', {'files': files})
 
-
-
-# @login_required
+@login_required
 def preview(request, file_id):
     file = get_object_or_404(FileModels, id=file_id)
     if file.file.name.lower().endswith('pdf'):
         return FileResponse(open(file.file.path, 'rb'), content_type='application/pdf')
-    elif file.file.name.lower().endswith(('jpg','jpeg', 'png', 'gif')):
-        print(file.file.url)
+    elif file.file.name.lower().endswith(('jpg', 'jpeg', 'png', 'gif')):
         return render(request, 'filesystem/preview.html', {'file': file})
-    
     else:
         return render(request, 'filesystem/preview.html', {'file': file})
-    
 
-
-# @login_required
+@login_required
 def open_page(request, file_id):
     file = get_object_or_404(FileModels, id=file_id)
     return render(request, 'filesystem/open_page.html', {'file': file})
